security default setting w/o deprecated
